Use `mock_platform` for event entity component tests instead of `hass.components`
