+++ tests/components/event/const.py
@@ -1,0 +1,3 @@
+"""Constants for the event entity component tests."""
+
+TEST_DOMAIN = "test"
