+++ tests/components/event/test_init.py
@@ -22,6 +22,8 @@ from homeassistant.helpers.restore_state import STORAGE_KEY as RESTORE_STATE_KEY
 from homeassistant.setup import async_setup_component
 from homeassistant.util import dt as dt_util
 
+from .const import TEST_DOMAIN
+
 from tests.common import (
     MockConfigEntry,
     MockModule,
@@ -33,8 +35,6 @@ from tests.common import (
     mock_restore_cache,
     mock_restore_cache_with_extra_data,
 )
-
-TEST_DOMAIN = "test"
 
 
 async def test_event() -> None:
@@ -96,7 +96,7 @@ async def test_event() -> None:
         event._trigger_event("unknown_event")
 
 
-@pytest.mark.usefixtures("enable_custom_integrations")
+@pytest.mark.usefixtures("enable_custom_integrations", "mock_event_platform")
 async def test_restore_state(hass: HomeAssistant) -> None:
     """Test we restore state integration."""
     mock_restore_cache_with_extra_data(
@@ -128,9 +128,6 @@ async def test_restore_state(hass: HomeAssistant) -> None:
         ),
     )
 
-    platform = getattr(hass.components, f"test.{DOMAIN}")
-    platform.init()
-
     assert await async_setup_component(hass, DOMAIN, {DOMAIN: {CONF_PLATFORM: "test"}})
     await hass.async_block_till_done()
 
@@ -142,7 +139,7 @@ async def test_restore_state(hass: HomeAssistant) -> None:
     assert state.attributes["hello"] == "world"
 
 
-@pytest.mark.usefixtures("enable_custom_integrations")
+@pytest.mark.usefixtures("enable_custom_integrations", "mock_event_platform")
 async def test_invalid_extra_restore_state(hass: HomeAssistant) -> None:
     """Test we restore state integration."""
     mock_restore_cache_with_extra_data(
@@ -163,9 +160,6 @@ async def test_invalid_extra_restore_state(hass: HomeAssistant) -> None:
         ),
     )
 
-    platform = getattr(hass.components, f"test.{DOMAIN}")
-    platform.init()
-
     assert await async_setup_component(hass, DOMAIN, {DOMAIN: {CONF_PLATFORM: "test"}})
     await hass.async_block_till_done()
 
@@ -177,7 +171,7 @@ async def test_invalid_extra_restore_state(hass: HomeAssistant) -> None:
     assert "hello" not in state.attributes
 
 
-@pytest.mark.usefixtures("enable_custom_integrations")
+@pytest.mark.usefixtures("enable_custom_integrations", "mock_event_platform")
 async def test_no_extra_restore_state(hass: HomeAssistant) -> None:
     """Test we restore state integration."""
     mock_restore_cache(
@@ -198,9 +192,6 @@ async def test_no_extra_restore_state(hass: HomeAssistant) -> None:
         ),
     )
 
-    platform = getattr(hass.components, f"test.{DOMAIN}")
-    platform.init()
-
     assert await async_setup_component(hass, DOMAIN, {DOMAIN: {CONF_PLATFORM: "test"}})
     await hass.async_block_till_done()
 
@@ -212,7 +203,7 @@ async def test_no_extra_restore_state(hass: HomeAssistant) -> None:
     assert "hello" not in state.attributes
 
 
-@pytest.mark.usefixtures("enable_custom_integrations")
+@pytest.mark.usefixtures("enable_custom_integrations", "mock_event_platform")
 async def test_saving_state(hass: HomeAssistant, hass_storage: dict[str, Any]) -> None:
     """Test we restore state integration."""
     restore_data = {"last_event_type": "double_press", "last_event_attributes": None}
@@ -229,9 +220,6 @@ async def test_saving_state(hass: HomeAssistant, hass_storage: dict[str, Any]) -
             ),
         ),
     )
-
-    platform = getattr(hass.components, f"test.{DOMAIN}")
-    platform.init()
 
     assert await async_setup_component(hass, DOMAIN, {DOMAIN: {CONF_PLATFORM: "test"}})
     await hass.async_block_till_done()
